#17 flake8 issues
