+++ docs/conf.py
@@ -14,6 +14,7 @@
 
 import os
 import sys
+from typing import Dict, Optional
 
 import tomlkit
 
@@ -22,7 +23,7 @@ sys.path.insert(0, os.path.abspath('..'))
 
 # -- Project information -----------------------------------------------------
 
-def _get_project_meta():
+def _get_project_meta() -> Dict[str, str]:
     with open('../pyproject.toml') as pyproject:
         file_contents = pyproject.read()
 
@@ -96,7 +97,7 @@ master_doc = 'index'
 #
 # This is also used if you do content translation via gettext catalogs.
 # Usually you set "language" from the command line for these cases.
-language = None
+language: Optional[str] = None
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
