+++ platonic/sqs/queue/__init__.py
@@ -1,8 +1,8 @@
-from platonic.sqs.queue.types import InternalType, ValueType
-from platonic.sqs.queue.input import SQSInputQueue
-from platonic.sqs.queue.output import SQSOutputQueue
-from platonic.sqs.queue.message import SQSMessage
 from platonic.sqs.queue.errors import (
     SQSMessageDoesNotExist,
     SQSQueueDoesNotExist,
 )
+from platonic.sqs.queue.input import SQSInputQueue
+from platonic.sqs.queue.message import SQSMessage
+from platonic.sqs.queue.output import SQSOutputQueue
+from platonic.sqs.queue.types import InternalType, ValueType
