+++ platonic/sqs/queue/sqs.py
@@ -7,11 +7,11 @@ from typecasts import Typecasts, casts
 
 from platonic.const import const
 
+# Max number of SQS messages receivable by single API call.
 MAX_NUMBER_OF_MESSAGES = 10
-"""Max number of SQS messages receivable by single API call."""
 
+# Message in its raw form must be shorter than this.
 MAX_MESSAGE_SIZE = 262144
-"""Message in its raw form must be shorter than this."""
 
 
 @dataclasses.dataclass
