+++ platonic/sqs/queue/types.py
@@ -1,4 +1,4 @@
-from typing import TypeVar, NewType
+from typing import NewType, TypeVar
 
 ValueType = TypeVar('ValueType')
 InternalType = NewType('InternalType', str)
