+++ pyproject.toml
@@ -56,3 +56,4 @@ tomlkit = "^0.5"
 boto3 = "^1.12.11"
 moto = "^1.3.14"
 contexttimer = "^0.3.3"
+isort = "<5.0.0"
