+++ setup.cfg
@@ -60,6 +60,10 @@ per-file-ignores =
     D101, D103,
     # And short variable names
     WPS110
+    # Lists multiplied
+    WPS435
+    # Found overused expression
+    WPS204
 
   __init__.py:
     # it is allowed to import things you're not using in the __init__.py file itself.
@@ -72,6 +76,13 @@ per-file-ignores =
   platonic/sqs/conversions/*.py:
     # Allow `value` as variable name
     WPS110
+
+  tests/test_queue/robot.py:
+    # Allow uppercase constants in an enum
+    WPS115
+
+    # Subclassing is ok
+    WPS600
 
 [isort]
 # isort configuration:
