+++ tests/test_queue/conftest.py
@@ -1,12 +1,18 @@
 import os
 
 import boto3
+import pytest
 from moto.sqs import mock_sqs
 from mypy_boto3_sqs import Client as SQSClient
-import pytest
+
+from tests.test_queue.robot import (
+    CommandReceiver,
+    CommandSender,
+    ReceiverAndSender,
+)
 
 
-@pytest.fixture(scope='function')
+@pytest.fixture()
 def mock_sqs_client():
     """Mocked AWS Credentials for moto."""
     os.environ['AWS_ACCESS_KEY_ID'] = 'testing'
@@ -16,3 +22,21 @@ def mock_sqs_client():
 
     with mock_sqs():
         yield boto3.client('sqs')
+
+
+@pytest.fixture()
+def receiver_and_sender(
+    mock_sqs_client: SQSClient,  # noqa: WPS442
+) -> ReceiverAndSender:
+    """Construct two connected SQS queues."""
+    sqs_queue_url = mock_sqs_client.create_queue(
+        QueueName='robot_commands',
+        Attributes={
+            'VisibilityTimeout': '3',
+        },
+    )['QueueUrl']
+
+    sender = CommandSender(url=sqs_queue_url)
+    receiver = CommandReceiver(url=sqs_queue_url)
+
+    return receiver, sender
