+++ tests/test_queue/test_message_size.py
@@ -1,9 +1,9 @@
 """Children are sending letters to Santa Claus. Sometimes those are too long."""
 import pytest
+from mypy_boto3_sqs import Client as SQSClient
 
 from platonic.queue import MessageTooLarge
-from platonic.sqs.queue import SQSOutputQueue, SQSInputQueue
-from mypy_boto3_sqs import Client as SQSClient
+from platonic.sqs.queue import SQSInputQueue, SQSOutputQueue
 
 
 class LetterSender(SQSOutputQueue[str]):
