+++ tests/test_queue/test_queue.py
@@ -5,34 +5,8 @@ import contexttimer
 import pytest
 from mypy_boto3_sqs import Client as SQSClient
 
-from platonic.queue import (
-    QueueDoesNotExist,
-    MessageDoesNotExist,
-    MessageReceiveTimeout,
-)
-from platonic.sqs.queue import SQSMessage
-from tests.test_queue.robot import (
-    Command, CommandSender, CommandReceiver,
-    ReceiverAndSender,
-)
-
-
-@pytest.fixture
-def receiver_and_sender(
-    mock_sqs_client: SQSClient,
-) -> ReceiverAndSender:
-    """Construct two connected SQS queues."""
-    sqs_queue_url = mock_sqs_client.create_queue(
-        QueueName='robot_commands',
-        Attributes={
-            'VisibilityTimeout': '3',
-        }
-    )['QueueUrl']
-
-    sender = CommandSender(url=sqs_queue_url)
-    receiver = CommandReceiver(url=sqs_queue_url)
-
-    return receiver, sender
+from platonic.queue import MessageReceiveTimeout, QueueDoesNotExist
+from tests.test_queue.robot import Command, CommandSender, ReceiverAndSender
 
 
 def test_non_existing_queue(mock_sqs_client: SQSClient):
@@ -65,35 +39,10 @@ def test_send_once_and_receive_twice(receiver_and_sender: ReceiverAndSender):
     with contexttimer.Timer() as timer:
         assert receiver.receive().value == Command.RIGHT
 
+        elapsed_time = timer.elapsed
+
     # The visibility timeout for the mock SQS queue is 3 seconds.
-    assert round(timer.elapsed) >= 3
-
-
-def test_send_and_acknowledge(receiver_and_sender: ReceiverAndSender):
-    """We receive a command, acknowledge it, and now queue is empty."""
-    receiver, sender = receiver_and_sender
-    sender.send(Command.JUMP)
-
-    jump_message = receiver.receive()
-    receiver.acknowledge(jump_message)
-
-    # Now the queue is empty. We will not be able to receive this message
-    # once again.
-    with pytest.raises(MessageReceiveTimeout):
-        receiver.receive_with_timeout(timeout=2)
-
-
-def test_acknowledge_fake_message(receiver_and_sender: ReceiverAndSender):
-    """Acknowledging a message that does not exist causes an exception."""
-    receiver, sender = receiver_and_sender
-
-    message = SQSMessage[Command](
-        value=Command.JUMP,
-        id='abc',
-    )
-
-    with pytest.raises(MessageDoesNotExist):
-        receiver.acknowledge(message)
+    assert round(elapsed_time) >= 3
 
 
 def test_put_many(receiver_and_sender: ReceiverAndSender):
@@ -155,27 +104,6 @@ def test_iterate_functional(receiver_and_sender: ReceiverAndSender):
     assert received_commands == sent_commands
 
 
-def test_acknowledgement(receiver_and_sender: ReceiverAndSender):
-    """
-    The acknowledgement() context manager automatically acknowledges messages.
-    """
-    receiver, sender = receiver_and_sender
-
-    sent_commands = [Command.RIGHT, Command.FORWARD, Command.LEFT, Command.JUMP]
-
-    sender.send_many(sent_commands)
-
-    received_commands = []
-    for message in receiver:
-        with receiver.acknowledgement(message):
-            received_commands.append(message.value)
-
-        if message.value == Command.JUMP:
-            break
-
-    assert received_commands == sent_commands
-
-
 def test_receive_with_timeout_on_empty_queue(
     receiver_and_sender: ReceiverAndSender,
 ):
@@ -186,4 +114,6 @@ def test_receive_with_timeout_on_empty_queue(
         with pytest.raises(MessageReceiveTimeout):
             receiver.receive_with_timeout(timeout=1)
 
-    assert round(timer.elapsed) >= 1, 'We have waited for 1 second at least.'
+        elapsed_time = timer.elapsed
+
+    assert round(elapsed_time) >= 1, 'We have waited for 1 second at least.'
